Create Live_Tracker_Amina.py
Just something I tried out to see if I could improve the project. It is a slightly different approach but I believe it is more accurate. There is a way to outline the human with incredible precision instead of using a box with this model. If that would be useful I could work on that too.